Move some low-level locking primitives to be Unix-only since they are no longer used on Windows.
